+++ lib/src/models/calendar_style_config.dart
@@ -2,9 +2,16 @@ import 'package:flutter/material.dart';
 
 class CalendarStyleConfig {
   const CalendarStyleConfig({
+    this.backgroundColor,
+    this.calendarRadius,
+    this.headerYearTextStyle,
+    this.headerMonthTextStyle,
+    this.headerIconColor,
+    this.materialWrapperRadius,
     this.regularDateDecoration,
     this.selectedDateDecoration,
     this.dateDisplayFieldDecoration,
+    this.dateDisplayFieldSelectedDecoration,
     this.regularDateTextStyle,
     this.selectedDateTextStyle,
     this.todayDateTextStyle,
@@ -13,9 +20,16 @@ class CalendarStyleConfig {
     this.dateDisplayFieldSelectedTextStyle,
   });
 
+  final Color? backgroundColor;
+  final double? calendarRadius;
+  final TextStyle? headerYearTextStyle;
+  final TextStyle? headerMonthTextStyle;
+  final Color? headerIconColor;
+  final double? materialWrapperRadius;
   final BoxDecoration? regularDateDecoration;
   final BoxDecoration? selectedDateDecoration;
   final BoxDecoration? dateDisplayFieldDecoration;
+  final BoxDecoration? dateDisplayFieldSelectedDecoration;
   final TextStyle? regularDateTextStyle;
   final TextStyle? selectedDateTextStyle;
   final TextStyle? todayDateTextStyle;
@@ -25,6 +39,18 @@ class CalendarStyleConfig {
 
   factory CalendarStyleConfig.defaultStyle() {
     return CalendarStyleConfig(
+      backgroundColor: Colors.white,
+      calendarRadius: 16,
+      headerYearTextStyle: const TextStyle(
+        fontWeight: FontWeight.w500,
+        fontSize: 14,
+      ),
+      headerMonthTextStyle: const TextStyle(
+        fontWeight: FontWeight.w500,
+        fontSize: 14,
+      ),
+      headerIconColor: Colors.black,
+      materialWrapperRadius: 46,
       regularDateDecoration: BoxDecoration(
         color: Colors.transparent,
         borderRadius: BorderRadius.circular(46),
@@ -35,6 +61,10 @@ class CalendarStyleConfig {
       ),
       dateDisplayFieldDecoration: BoxDecoration(
         border: Border.all(color: const Color(0xffE2E8F0)),
+        borderRadius: BorderRadius.circular(12),
+      ),
+      dateDisplayFieldSelectedDecoration: BoxDecoration(
+        border: Border.all(color: const Color(0xff556EE6)),
         borderRadius: BorderRadius.circular(12),
       ),
       regularDateTextStyle: const TextStyle(
@@ -61,10 +91,57 @@ class CalendarStyleConfig {
       ),
     );
   }
+
+  CalendarStyleConfig copyWith({
+    Color? backgroundColor,
+    double? calendarRadius,
+    TextStyle? headerYearTextStyle,
+    TextStyle? headerMonthTextStyle,
+    Color? headerIconColor,
+    double? materialWrapperRadius,
+    BoxDecoration? regularDateDecoration,
+    BoxDecoration? selectedDateDecoration,
+    BoxDecoration? dateDisplayFieldDecoration,
+    BoxDecoration? dateDisplayFieldSelectedDecoration,
+    TextStyle? regularDateTextStyle,
+    TextStyle? selectedDateTextStyle,
+    TextStyle? todayDateTextStyle,
+    TextStyle? pastDateTextStyle,
+    TextStyle? dateDisplayFieldTextStyle,
+    TextStyle? dateDisplayFieldSelectedTextStyle,
+  }) {
+    return CalendarStyleConfig(
+      backgroundColor: backgroundColor ?? this.backgroundColor,
+      calendarRadius: calendarRadius ?? this.calendarRadius,
+      headerYearTextStyle: headerYearTextStyle ?? this.headerYearTextStyle,
+      headerMonthTextStyle: headerMonthTextStyle ?? this.headerMonthTextStyle,
+      headerIconColor: headerIconColor ?? this.headerIconColor,
+      materialWrapperRadius: materialWrapperRadius ?? this.materialWrapperRadius,
+      regularDateDecoration: regularDateDecoration ?? this.regularDateDecoration,
+      selectedDateDecoration: selectedDateDecoration ?? this.selectedDateDecoration,
+      dateDisplayFieldDecoration: dateDisplayFieldDecoration ?? this.dateDisplayFieldDecoration,
+      dateDisplayFieldSelectedDecoration: dateDisplayFieldSelectedDecoration ?? this.dateDisplayFieldSelectedDecoration,
+      regularDateTextStyle: regularDateTextStyle ?? this.regularDateTextStyle,
+      selectedDateTextStyle: selectedDateTextStyle ?? this.selectedDateTextStyle,
+      todayDateTextStyle: todayDateTextStyle ?? this.todayDateTextStyle,
+      pastDateTextStyle: pastDateTextStyle ?? this.pastDateTextStyle,
+      dateDisplayFieldTextStyle: dateDisplayFieldTextStyle ?? this.dateDisplayFieldTextStyle,
+      dateDisplayFieldSelectedTextStyle: dateDisplayFieldSelectedTextStyle ?? this.dateDisplayFieldSelectedTextStyle,
+    );
+  }
 }
 
 class RangeCalendarStyleConfig {
   const RangeCalendarStyleConfig({
+    this.backgroundColor,
+    this.calendarRadius,
+    this.headerYearTextStyle,
+    this.headerMonthTextStyle,
+    this.headerIconColor,
+    this.materialWrapperRadius,
+    this.rangeSegmentLeftRadius,
+    this.rangeSegmentRightRadius,
+    this.rangeSegmentBothSidesRadius,
     this.startDateDecoration,
     this.endDateDecoration,
     this.inRangeDateDecoration,
@@ -83,6 +160,15 @@ class RangeCalendarStyleConfig {
     this.dateDisplayFieldSelectedTextStyle,
   });
 
+  final Color? backgroundColor;
+  final double? calendarRadius;
+  final TextStyle? headerYearTextStyle;
+  final TextStyle? headerMonthTextStyle;
+  final Color? headerIconColor;
+  final double? materialWrapperRadius;
+  final double? rangeSegmentLeftRadius;
+  final double? rangeSegmentRightRadius;
+  final double? rangeSegmentBothSidesRadius;
   final BoxDecoration? startDateDecoration;
   final BoxDecoration? endDateDecoration;
   final BoxDecoration? inRangeDateDecoration;
@@ -102,6 +188,21 @@ class RangeCalendarStyleConfig {
 
   factory RangeCalendarStyleConfig.defaultStyle() {
     return RangeCalendarStyleConfig(
+      backgroundColor: Colors.white,
+      calendarRadius: 16,
+      headerYearTextStyle: const TextStyle(
+        fontWeight: FontWeight.w500,
+        fontSize: 14,
+      ),
+      headerMonthTextStyle: const TextStyle(
+        fontWeight: FontWeight.w500,
+        fontSize: 14,
+      ),
+      headerIconColor: Colors.black,
+      materialWrapperRadius: 46,
+      rangeSegmentLeftRadius: 50,
+      rangeSegmentRightRadius: 50,
+      rangeSegmentBothSidesRadius: 50,
       startDateDecoration: BoxDecoration(
         color: const Color(0xffE8EAF6),
         border: Border.all(color: const Color(0xff556EE6), width: 2),
@@ -156,6 +257,62 @@ class RangeCalendarStyleConfig {
       dateDisplayFieldSelectedTextStyle: const TextStyle(
         color: Color(0xff000000),
       ),
+    );
+  }
+
+  RangeCalendarStyleConfig copyWith({
+    Color? backgroundColor,
+    double? calendarRadius,
+    TextStyle? headerYearTextStyle,
+    TextStyle? headerMonthTextStyle,
+    Color? headerIconColor,
+    double? materialWrapperRadius,
+    double? rangeSegmentLeftRadius,
+    double? rangeSegmentRightRadius,
+    double? rangeSegmentBothSidesRadius,
+    BoxDecoration? startDateDecoration,
+    BoxDecoration? endDateDecoration,
+    BoxDecoration? inRangeDateDecoration,
+    Color? rangeSegmentColor,
+    BorderRadius? startDateRadius,
+    BorderRadius? endDateRadius,
+    BoxDecoration? dateDisplayFieldDecoration,
+    BoxDecoration? dateDisplayFieldSelectedDecoration,
+    TextStyle? regularDateTextStyle,
+    TextStyle? startDateTextStyle,
+    TextStyle? endDateTextStyle,
+    TextStyle? inRangeDateTextStyle,
+    TextStyle? todayDateTextStyle,
+    TextStyle? pastDateTextStyle,
+    TextStyle? dateDisplayFieldTextStyle,
+    TextStyle? dateDisplayFieldSelectedTextStyle,
+  }) {
+    return RangeCalendarStyleConfig(
+      backgroundColor: backgroundColor ?? this.backgroundColor,
+      calendarRadius: calendarRadius ?? this.calendarRadius,
+      headerYearTextStyle: headerYearTextStyle ?? this.headerYearTextStyle,
+      headerMonthTextStyle: headerMonthTextStyle ?? this.headerMonthTextStyle,
+      headerIconColor: headerIconColor ?? this.headerIconColor,
+      materialWrapperRadius: materialWrapperRadius ?? this.materialWrapperRadius,
+      rangeSegmentLeftRadius: rangeSegmentLeftRadius ?? this.rangeSegmentLeftRadius,
+      rangeSegmentRightRadius: rangeSegmentRightRadius ?? this.rangeSegmentRightRadius,
+      rangeSegmentBothSidesRadius: rangeSegmentBothSidesRadius ?? this.rangeSegmentBothSidesRadius,
+      startDateDecoration: startDateDecoration ?? this.startDateDecoration,
+      endDateDecoration: endDateDecoration ?? this.endDateDecoration,
+      inRangeDateDecoration: inRangeDateDecoration ?? this.inRangeDateDecoration,
+      rangeSegmentColor: rangeSegmentColor ?? this.rangeSegmentColor,
+      startDateRadius: startDateRadius ?? this.startDateRadius,
+      endDateRadius: endDateRadius ?? this.endDateRadius,
+      dateDisplayFieldDecoration: dateDisplayFieldDecoration ?? this.dateDisplayFieldDecoration,
+      dateDisplayFieldSelectedDecoration: dateDisplayFieldSelectedDecoration ?? this.dateDisplayFieldSelectedDecoration,
+      regularDateTextStyle: regularDateTextStyle ?? this.regularDateTextStyle,
+      startDateTextStyle: startDateTextStyle ?? this.startDateTextStyle,
+      endDateTextStyle: endDateTextStyle ?? this.endDateTextStyle,
+      inRangeDateTextStyle: inRangeDateTextStyle ?? this.inRangeDateTextStyle,
+      todayDateTextStyle: todayDateTextStyle ?? this.todayDateTextStyle,
+      pastDateTextStyle: pastDateTextStyle ?? this.pastDateTextStyle,
+      dateDisplayFieldTextStyle: dateDisplayFieldTextStyle ?? this.dateDisplayFieldTextStyle,
+      dateDisplayFieldSelectedTextStyle: dateDisplayFieldSelectedTextStyle ?? this.dateDisplayFieldSelectedTextStyle,
     );
   }
 }