+++ CHANGELOG.md
@@ -1,2 +1,6 @@
-## [1.0.0]
+### [0.0.2]
+- style config expansion (added copyWith for config classes and customization for Calendar()
+and RangeCalendar() background color, radius, header text style, border radius of date range
+in RangeCalendar())
+### [0.0.1]
 - init